requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.11.12
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
Django==4.2.26
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.9.1
lxml==6.0.2
openai==2.2.0
packaging==25.0
pdfminer.six==20250324
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyMuPDF==1.24.11
python-docx==1.1.2
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
